+++ src/main.js
@@ -121,12 +121,12 @@ const exitPyProc = (serverConfiguration) => {
 
 app.on("ready", () => {
   createPyProc(PysodaConfiguration);
-  createPyProc(pyFlaskConfiguration);
+  // createPyProc(pyFlaskConfiguration);
 });
 
 app.on("will-quit", () => {
   exitPyProc(PysodaConfiguration);
-  exitPyProc(pyFlaskConfiguration);
+  // exitPyProc(pyFlaskConfiguration);
 });
 
 /*************************************************************
@@ -180,7 +180,8 @@ function initialize() {
       } else {
         var first_launch = nodeStorage.getItem("firstlaunch");
         nodeStorage.setItem("firstlaunch", true);
-        exitPyProc();
+        exitPyProc(PysodaConfiguration);
+        // to do: exitPyProc(pyFlaskConfiguration);
         app.exit();
       }
     });