+++ src/scripts/metadata-files/manifest.js
@@ -1046,7 +1046,7 @@ function validateSPARCdataset() {
   localDatasetFolderPath = $("#input-manifest-local-folder-dataset").attr(
     "placeholder"
   );
-  valid_dataset = verify_sparc_folder(localDatasetFolderPath);
+  valid_dataset = verify_sparc_folder(localDatasetFolderPath, "local");
   if (valid_dataset == true) {
     let action = "";
     irregularFolderArray = [];
@@ -1334,7 +1334,6 @@ function createManifestLocally(type, editBoolean, originalDataset) {
                   },
                 }).then((result) => {});
               } else {
-                // console.log(res)
                 Swal.fire({
                   title: "Successfully generated!",
                   heightAuto: false,